BasketApp finally done
This one took me days. I decided to start this repository only because of this assignment.
I really hate the difficulty of managing dependencies and their versions in Android Studio.
The app is supposed to be for persisting Basketball matches. For the time being (and probably forever) the only entities will be Partido (match) and Cuarto (quarter).
For further understanding of the project, look in the AppBaloncestoFran\app\src\main\java\com\example\appbaloncestofran\entities folder and inspect both files.